moved to the right component
