+++ frontend/src/routes/SingleListing.tsx
@@ -17,7 +17,7 @@ const SingleListing = () => {
 
       <div className="flex">
         <PropertyPicsCollage />
-        <div>
+        <div className="space-y-8">
           <Calendar />
           <MoveInCalc />
           <GetInTouch />
